Substantial refactor, fixes, and adjustment to align with vision
