+++ backend_ts/src/index.ts
@@ -1,14 +1,13 @@
 import { pathfinder } from "mineflayer-pathfinder";
 import type { Bot, BotOptions } from "mineflayer";
-import { PathfinderAbstraction } from "./movement/movement";
-import { PathfinderAbstractionOptions } from "./movement/types";
-import { SemanticWorld } from "./world/worldState1";
-import { SemanticOptions } from "./types";
+import { PathfinderAbstraction } from "./pathfinding"
+import { EnvState } from "./envState";
+import { SurroundingsOptions } from "./types";
 
 declare module "mineflayer" {
   interface Bot {
     pathfinderAbstract: PathfinderAbstraction;
-    semanticWorld: SemanticWorld;
+    envState: EnvState;
   }
 }
 
@@ -16,10 +15,10 @@ declare module "mineflayer" {
 
 
 
-export function createPlugin(opts: SemanticOptions) {
+export function createPlugin(opts: SurroundingsOptions) {
   return (bot: Bot, botOptions: BotOptions) => {
     bot.pathfinderAbstract = new PathfinderAbstraction(bot);
-    bot.semanticWorld = new SemanticWorld(bot, opts);
+    bot.envState = new EnvState(bot, opts);
     if (!bot.hasPlugin(pathfinder)) bot.loadPlugin(pathfinder);
   };
 }
