+++ backend_ts/src/types.ts
@@ -1,4 +1,9 @@
-export type SemanticOptions = {
-  immediateRadius: number;
-  nearbyRadius: number;
+import { PCChunk } from "prismarine-chunk";
+
+
+export type SurroundingsOptions = {
+  immediateSurroundingsRadius: number;
+  distantSurroundingsRadius: number;
 };
+
+export type PCChunkCoordinateAndColumn = { chunkX: number; chunkZ: number; column: PCChunk };
