--- backend_ts/src/world/types.ts
@@ -1,11 +1,0 @@
-import { PCChunk } from "prismarine-chunk";
-
-export interface NearbySurroundingsConfig {
-  blockRadius: number;
-}
-
-export type PCChunkCoordinateAndColumn = { chunkX: number; chunkZ: number; column: PCChunk };
-
-export interface ImmediateSurroundingsConfig {
-  radius: number;
-}
